Moved WebGL2 context initial configuration to own module
